refactor: organize demo
