+++ packages/vite-node-miniflare/demo/src/app.tsx
