+++ packages/vite-node-miniflare/demo/src/client.ts
