+++ packages/vite-node-miniflare/demo/src/component.tsx
